Resolve #24 - Change copyright to include Aguesoftguar project. Reorder project packages and rename classes
