+++ app/src/main/AndroidManifest.xml
@@ -7,14 +7,14 @@
     <uses-permission android:name="android.permission.WAKE_LOCK" />
 
     <application
-        android:name=".main.MedAlarmApp"
+        android:name=".medicines.MedAlarmApp"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".main.MainActivity"
+            android:name=".medicines.MedicinesActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
@@ -24,7 +24,7 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".main.addeditmedicine.AddEditMedicineActivity" />
+        <activity android:name=".addeditmedicine.AddEditMedicineActivity" />
 
     </application>
