+++ app/src/main/java/com/aguesoftguar/medalarm/addeditmedicine/AddEditMedicineActivity.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016, The Android Open Source Project
+ * Copyright 2016, Agüesoftgüar
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,15 +14,15 @@
  * limitations under the License.
  */
 
-package com.aguesoftguar.medalarm.main.addeditmedicine;
+package com.aguesoftguar.medalarm.addeditmedicine;
 
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.View;
 
-import com.aguesoftguar.medalarm.ActivityUtils;
 import com.aguesoftguar.medalarm.R;
+import com.aguesoftguar.medalarm.util.ActivityUtils;
 
 /**
  * Displays an add or edit medicines screen.
