+++ app/src/main/java/com/aguesoftguar/medalarm/util/ActivityUtils.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016, The Android Open Source Project
+ * Copyright 2016, Agüesoftgüar
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.aguesoftguar.medalarm;
+package com.aguesoftguar.medalarm.util;
 
 import android.support.annotation.NonNull;
 import android.support.v4.app.Fragment;
