+++ app/src/main/res/layout/fragment_main.xml
@@ -9,7 +9,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.aguesoftguar.medalarm.main.MainActivity"
+    tools:context="com.aguesoftguar.medalarm.medicines.MedicinesActivity"
     tools:showIn="@layout/activity_main">
 
     <TextView
